+++ imml/decomposition/jnmf.py
@@ -25,7 +25,7 @@ if rmodule_installed:
 class JNMF(TransformerMixin, BaseEstimator):
     r"""
     Joint Non-Negative Matrix Factorization (JNMF).
-    [#jnmfpaper1]_ [#jnmfpaper2]_ [#jnmfpaper3]_ [#jnmfpaper4]_ [#jnmfpaper5]_ [#jnmfpaper6]_ [#jnmfcode1]_ [#jnmfcode2]_
+    [#jnmfpaper1]_ [#jnmfpaper2]_ [#jnmfcode1]_ [#jnmfcode2]_
 
     JNMF decompose the matrices to low-dimensional factor matrices.
 
@@ -92,18 +92,9 @@ class JNMF(TransformerMixin, BaseEstimator):
     ----------
     .. [#jnmfpaper1] Tsuyuzaki et al., (2023). nnTensor: An R package for non-negative matrix/tensor decomposition.
                      Journal of Open Source Software, 8(84), 5015, https://doi.org/10.21105/joss.05015
-    .. [#jnmfpaper2] Liviu Badea, (2008) Extracting Gene Expression Profiles Common to Colon and Pancreatic
-                     Adenocarcinoma using Simultaneous nonnegative matrix factorization. Pacific Symposium on
-                     Biocomputing 13:279-290.
-    .. [#jnmfpaper3] Shihua Zhang, et al. (2012) Discovery of multi-dimensional modules by integrative analysis of
-                     cancer genomic data. Nucleic Acids Research 40(19), 9379-9391.
-    .. [#jnmfpaper4] Zi Yang, et al. (2016) A non-negative matrix factorization method for detecting modules in
+    .. [#jnmfpaper2] Zi Yang, et al. (2016) A non-negative matrix factorization method for detecting modules in
                      heterogeneous omics multi-modal data, Bioinformatics 32(1), 1-8.
-    .. [#jnmfpaper5] Y. Kenan Yilmaz et al., (2010) Probabilistic Latent Tensor Factorization, International Conference
-                     on Latent Variable Analysis and Signal Separation 346-353.
-    .. [#jnmfpaper6] N. Fujita et al., (2018) Biomarker discovery by integrated joint non-negative matrix factorization
-                     and pathway signature analyses, Scientific Report.
-    .. [#jnmfcode1] https://rdrr.io/cran/nnTensor/man/JNMF.html
+    .. [#jnmfcode1] https://cran.r-project.org/web/packages/nnTensor/vignettes/nnTensor-2.html
     .. [#jnmfcode2] https://github.com/rikenbit/nnTensor
 
     See Also